+++ packages/wallet/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Pass optional `getCanonicalProfileId` and `metaMetricsFlags` through `RemoteFeatureFlagController` wallet initialization for threshold flag segmentation ([#9325](https://github.com/MetaMask/core/pull/9325))
+
 ## [7.0.0]
 
 ### Added